+++ Functions/WaitbarConsole.py
